+++ elevenlabs_rs/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- **Breaking:**`GetVoiceQuery` as now deprecated
 ## [0.5.1] - 2025-02-28
 ### Added
 - moved `SharedVoice` back to `voice_library` module